requirements.txt
fastapi[standard]
sqlalchemy
passlib
bcrypt==4.0.1
python-jose[cryptography]